Move opendata_scraper, update references
