--- common/__init__.py
@@ -1,2 +1,0 @@
-from .base_scrapers.opendata import opendata_scraper
-from .base_scrapers.opendata import opendata_scraper2
