--- common/base_scrapers/__init__.py
@@ -1,2 +1,0 @@
-from .opendata import opendata_scraper
-from .opendata import opendata_scraper2
