+++ scrapers/CA/los_angeles_county/municipality/los_angeles/la_opendata.py
@@ -5,7 +5,7 @@ from pathlib import Path
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
 
-from common import opendata_scraper2
+from scrapers.data_portals.opendata.opendata_scraper_2 import opendata_scraper2
 
 save_url = [
     ["crime_data/2020-present/", "https://data.lacity.org/resource/2nrs-mtv8.csv"],
