+++ scrapers/FL/pinellas_county/municipal/st_petersburg/st_petersburg_opendata.py
@@ -5,7 +5,7 @@ from pathlib import Path
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
 
-from common import opendata_scraper2
+from scrapers.data_portals.opendata.opendata_scraper_2 import opendata_scraper2
 
 save_url = [
     ["cfs/", "https://stat.stpete.org/resource/2eks-pg5j.csv"],
