+++ scrapers/OH/hamilton_county/municipal/cincinnati/cincinnati_opendata.py
@@ -4,7 +4,7 @@ from pathlib import Path
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
 
-from common import opendata_scraper2
+from scrapers.data_portals.opendata.opendata_scraper_2 import opendata_scraper2
 
 save_url = [
     ["cfs/", "https://data.cincinnati-oh.gov/resource/gexm-h6bt.csv"],
