+++ scrapers/TX/travis_county/municipal/austin/austin_opendata.py
@@ -5,7 +5,7 @@ from pathlib import Path
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
 
-from common import opendata_scraper2
+from scrapers.data_portals.opendata.opendata_scraper_2 import opendata_scraper2
 
 # Change to what you need (remove what you don't)
 save_url = [
