+++ scrapers/WA/king_county/municipal/seattle/archive_seattle_opendata.py
@@ -5,7 +5,7 @@ from pathlib import Path
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
 
-from common import opendata_scraper2
+from scrapers.data_portals.opendata.opendata_scraper_2 import opendata_scraper2
 
 save_url = [
     ["pdrs/2016-2020/", "https://data.seattle.gov/resource/8fwq-jcnn.csv"],
