+++ scrapers/data_portals/__init__.py
@@ -1,5 +1,5 @@
 from .crimegraphics import crimegraphics_clery
 from .crimegraphics import crimegraphics_arrest
 from .crimegraphics import crimegraphics_bulletin
-#from .opendata import opendata_scraper
-#from .opendata import opendata_scraper2
+from .opendata import opendata_scraper
+from .opendata import opendata_scraper2
