+++ utils/setup_gui/Base_Scripts/Scrapers/opendata/opendata_scraper.py
@@ -5,7 +5,7 @@ from pathlib import Path
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
 
-from common import opendata_scraper2
+from scrapers.data_portals.opendata.opendata_scraper_2 import opendata_scraper2
 
 # This script was generated by the setup gui
 save_url = []
